Finish RESTAPI

+++ database/migrations/2023_07_09_204532_create_global_units_table.php
@@ -9,7 +9,7 @@
             Schema::create('global_units', function (Blueprint $table) {
                 $table->id();
                 $table->string('name');
-                $table->float('converter');
+                $table->decimal('converter');
                 $table->boolean('default');
                 $table->timestamps();
             });
